+++ crates/analyzer/src/api.rs
