+++ crates/analyzer/src/app.rs
