+++ crates/analyzer/src/check.rs
