+++ crates/analyzer/src/error.rs
