+++ crates/analyzer/src/log.rs
