+++ crates/analyzer/src/rpm.rs
