+++ crates/analyzer/src/rules/decision.rs
