+++ crates/analyzer/src/rules/file_type.rs
