+++ crates/analyzer/src/rules/mod.rs
