+++ crates/analyzer/src/rules/object.rs
