+++ crates/analyzer/src/rules/permission.rs
