+++ crates/analyzer/src/rules/rule.rs
