+++ crates/analyzer/src/rules/set.rs
