+++ crates/analyzer/src/rules/subject.rs
