+++ crates/analyzer/src/trust.rs
