--- python/tests/data/one.trust
@@ -1,1 +1,0 @@
-/bin/ls 133792 0d06f9724af41b13cdacea133530b9129a48450230feef9632d53d5bbb837c8c
