--- src/lib.rs
@@ -1,14 +1,0 @@
-pub mod api;
-pub mod app;
-pub mod cfg;
-pub mod check;
-pub mod error;
-pub mod fapolicyd;
-pub mod log;
-pub mod rpm;
-pub mod rules;
-pub mod sha;
-pub mod svc;
-pub mod sys;
-pub mod trust;
-pub mod users;
